added dll compiled sandcat for in-memory only
